backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
openai==1.51.2
google-cloud-bigquery==3.25.0
langgraph==0.2.34
langchain-core==0.3.17
pandas==2.2.3
fastavro==1.9.7
pyhive==0.7.0
psycopg2-binary==2.9.10
cx-Oracle==8.3.0
